refactor: mark constants file as private
