+++ fs/findNearestFile.test.ts
@@ -1,6 +1,6 @@
 import { resolve } from "../deps/path.ts";
 import { assertEquals, assertRejects, describe, it } from "../deps/testing.ts";
-import { FIXTURE_DIR } from "../constants.ts";
+import { FIXTURE_DIR } from "../_constants.ts";
 import { findNearestFile } from "./findNearestFile.ts";
 
 const A_DIR = resolve(FIXTURE_DIR, "a");
